cleanup based on upstream changes

+++ link.sh
@@ -8,10 +8,12 @@ mv /etc/munin/openquery/numa_ /etc/munin/openquery/numa_.bak
 wget --no-check-certificate https://raw.githubusercontent.com/munin-monitoring/munin/devel/plugins/node.d/mysql_ -O /etc/munin/openquery/mysql_ || exit 1
 wget --no-check-certificate https://raw.githubusercontent.com/munin-monitoring/munin/devel/plugins/node.d.linux/numa_ -O /etc/munin/openquery/numa_ || exit 1
 
-cp /usr/sbin/munin-asyncd /usr/sbin/munin-asyncd.orig
-wget --no-check-certificate https://raw.githubusercontent.com/munin-monitoring/munin/devel/node/_bin/munin-asyncd.in -O - | sed -e 's:@@PERL@@:/usr/bin/perl:' -e 's:@@SPOOLDIR@@:/var/lib/munin/spool:' > /usr/sbin/munin-asyncd || exit 1
+if [ ! -f /usr/sbin/munin-asyncd.orig ]; then
+    cp /usr/sbin/munin-asyncd /usr/sbin/munin-asyncd.orig
+    wget --no-check-certificate https://raw.githubusercontent.com/munin-monitoring/munin/devel/node/_bin/munin-asyncd -O /usr/sbin/munin-asyncd || exit 1
+    chown root: /usr/sbin/munin-asyncd
+fi
 
-chown root: /usr/sbin/munin-asyncd
 chmod a+x /usr/sbin/munin-asyncd /etc/munin/openquery/mysql_ /etc/munin/openquery/numa_
 
 munin-node-configure  --libdir /etc/munin/openquery/ --shell | sh
@@ -28,6 +30,6 @@ if selinuxenabled; then
   popd
 fi
 
-service munin-node restart || /etc/init.d/munin-node restart
+service munin-node restart
 [ -e /etc/init.d/munin-asyncd ] && service munin-asyncd restart
 [ -e /etc/init.d/munin-async  ] && service munin-async  restart
